+++ tests/indicator_matrix/test_indicator_matrix.py
@@ -20,7 +20,7 @@ import chispa as ch
 import pytest
 from pyspark.sql import types as T
 
-from scalelink.indicator_matrix import functions as im
+from scalelink.indicator_matrix import indicator_matrix as im
 
 
 def test_calculate_agreement_states(spark):