+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [semantic versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Indicator matrix functions and unit tests.
+- Utility functions and their unit tests.
 - Matrix A* functions and their unit tests.
 
 ### Changed